Made FloorController. Added support for fixed rooms
FloorController is a class that allows floors to be created and managed
as a single entity. Largely, it just offloads the work previously done
in HouseManager. This frees up HouseManager to do things like make
decisions on a  House wide scale. This is done with the future scenario
editor in mind. There is now functionality in place to specify initial
floor/house configurations and procedurally branch paths from there. At
this point, the path generation still needs some work to be optimal, but
it is generically functional.